% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\docType{package}
\name{summariser}
\alias{summariser}
\alias{summariser-package}
\title{summariser: an R package for easily calculating basic summary statistics}
\description{
The summariser package provides two important functions: \code{summary_stat#' s} and \code{summary_plot}
}
