% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{summary_plot}
\usage{
summary_plot(data, x, colour = NULL)
}
\arguments{
\item{data}{a data frame produced by \code{summary_stats}}

\item{x}{a factor variable to plot on the x axis}

\item{colour}{a factor variable to colour points by}
}
\description{
the output of a \code{summary_stats} call is input into a
 \code{summary_plot} call to produce a graph of the calculated means and
 confidence intervals.
}
\examples{
data(iris)
sum1 <- summary_stats(iris, measure = "Sepal.Length", Species)
summary_plot(sum1, x="Species", colour="Species")
}
