% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate summary statistics on a data frame}
\usage{
<<<<<<< HEAD:man/summary_stats.Rd
summary_stats(data, measure, ...)
=======
summary_stats(data, measure, type)
>>>>>>> dplyr_v7:man/summary_stats.Rd
}
\arguments{
\item{data}{a data frame}

\item{measure}{a numeric variable. Response variable - summary statistics
will be returned for this variable}
}
\description{
Functions from dplyr are used to automate the process of
 calculating basic summary statistics on a data frame. Returned
 statistics include mean, standard deviation, standard error, count,
 and 95% confidence intervals. Functions are given for calculating
 confidence intervals from a normal distribution (summary_stats) and
 from a t-distribution (summary_stats.t)
}
\examples{
data(iris)
  summary_stats(iris, measure = Sepal.Length)
  summary_stats(iris, measure = Sepal.Length, Species)
}
